+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -148,8 +148,8 @@ New checks
 
   Finds nondeterministic usages of pointers in unordered containers.
 
-- New :doc:`bugprone-reset-call
-  <clang-tidy/checks/bugprone/reset-call>` check.
+- New :doc:`bugprone-smartptr-reset-ambiguous-call
+  <clang-tidy/checks/bugprone/smartptr-reset-ambiguous-call>` check.
 
   Finds potentially erroneous calls to ``reset()`` method on smart pointers when
   the pointee type also has a ``reset()`` method.